Fix registration test cases
